Makefile: Make CC and CCFLAGS Configurable
Use weak assignment for CCFLAGS so it can respects predefined flags.

Make CC variable and therefore the compiler configurable by checking if
it is set already by using the origin function.
This is helpful to change the compiler version on the fly or to use a
specific compiler for a different platform.

Signed-off-by: Jens Erdmann <jens.erdmann@web.de>